pyproject.toml
[project]
name = "metrica"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "argon2-cffi>=25.1.0",
    "asyncpg>=0.31.0",
    "fastapi[all]>=0.129.0",
    "greenlet>=3.3.1",
    "httpx>=0.28.1",
    "passlib>=1.7.4",
    "pydantic-settings>=2.11.0",
    "python-jose>=3.5.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.46",
]
